requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# OpenAI Integration
openai==1.3.0

# Git Analysis
GitPython==3.1.30
pydriller==2.5

# Document Parsing
python-docx==1.1.0
python-pptx==0.6.23
pdfminer.six==20221105
reportlab==4.0.7

# Database
sqlalchemy==2.0.23

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Utilities
python-multipart==0.0.6
colorama==0.4.6

# Video Analysis
moviepy==1.0.3
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0

# Code Analysis
pygments>=2.16.0  
toml>=0.10.2  
requirements-parser>=0.5.0 

pytest
pandas
moviepy==1.0.3
ffmpeg-python
coverage

# Image Processing
Pillow==10.1.0
numpy==1.26.2
opencv-python==4.8.1.78
pytesseract==0.3.10
scikit-image==0.22.0
pyzbar==0.1.9
